RAC-79 rename: dto 패키지를 application 패키지 아래로 위치 변경
