+++ src/main/java/com/postgraduate/domain/auth/application/dto/KakaoLoginRequest.java
@@ -1,4 +1,4 @@
-package com.postgraduate.domain.auth.presentation.dto;
+package com.postgraduate.domain.auth.application.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
